Whitelist specific metadata folders
Narrowed down the metadata generator to only process the actual RPM folders, since that's the only interesting aspect when sharing archives with other computers.

It was unintentionally generating metadata for `.git` and other unrelated files, which was an oversight after moving the Archived Driver RPMs to a separate repository (which added all of those new Git files).