Whitelist mirror folders when pruning old packages
This was the second and final oversight when moving the RPM archives from being a plain folder to a sub-repository.

The (secret, internal) package pruning command has now been fixed to only delete old content from the RPM mirror folders, as intended. It was unintentionally traversing the sub-repository's `.git` folders.

Both of these minor patches were a result of forgetting that the metadata/pruning functions originally assumed that the RPM-mirror folder contained nothing except RPMs. But when it was time for release, it was decided to split the driver archives (official mirror) to a separate sub-repository due to GitHub's hosting limits for large files.

These small issues only affected people who plan to maintain the driver archives (such as myself, or anyone who decides to self-host), and would only affect newly generated mirrors. Nobody is affected, since these very minor issues were caught so close after the release, and it didn't affect the original mirror provided at the release. Everything works as it should!

Enjoy!