feat(restore-direct): Detect latest NVIDIA utility package revisions
This improves the "restore-direct" command:

Utility NVIDIA packages may have older revisions than the main driver, meaning that it was impossible to find the RPM files for the utilities in some situations, which made the installer give up due to "missing packages".

Therefore, I've implemented a new algorithm to automatically detect the latest revision of the utility packages which works with the desired NVIDIA driver version.