docs: Clarify supported Fedora releases and refine user instructions
This tool supports all current and future versions of Fedora, which should be mentioned in the README. Users shouldn't worry if this repository is several years old, since it will still work reliably.

A great example of this fact: The tool didn't receive any updates between 2022-2005 (3 years), but still worked perfectly in 2025. The main update in 2025 simply added DNF5 support, to handle its slightly different argument syntax compared to DNF4.

Also took the opportunity to improve some usability aspects, such as making it possible to specify the optional configuration values via environment variables, and a few enhancements to the "restore-direct" tool mode.

Looking forward to the next 30 years of not needing any updates!